[fix bug 2157411]
Use a background image (and text) instead of a hard-coded content image for
displaying item loading progress.
Added ITEM_LOADING_CONTENT string for configuring the message to be displayed
when an item is being loaded dynamically.
Modified ITEM_LOADING config to use the CSS class for the background spinner
image.