Added static constants to hold classNames and markup.
Renamed ariaText to ariaTemplate, to make it clear it is a
template to build the aria-valuetext attribute from.

Added attributes captionTemplate and captionEl.
captionTemplate, as ariaTemplate, will be used to overlay
a textual representation of the bar's value.
Defaults to nothing.

captionEl will be initialized to a div element above both
the bar and the mask (it won't be covered by the mask).
This element will normally hold the caption produced based on captionTemplate
It can be set to any other element or reference to the id of an element
to show that template elsewhere.

Dropped skin since I still have nothing that matches the sam skin