Added more to the unit testing and changed a few things
The redraw method reads minValue and maxValue so
it cannot be called from 'method' Attribute function
because they are not set yet. Redrawing is done when listening
to the  minValue and maxValue afterChange events, as it was initially.

Redraw already calls _recalculateConstants, there were redundant calls

Run the YuiDoc generator and noticed some bad descriptions, which I changed.