Cleaned up several images that were not being used
Fixed positioning of cuadrants by doing absoulte positioning as suggested by Jenny

Change examples:
	I now use one star or rankings, it tiles
        VUmeter is now contained within frame.  It was a matter of styling the container
	VUmeter moves with a timer

Added tests for animation (discovered a bug, good!)

Fixed bug: The bar initially is 100% wide and 100% tall,
since you don't know yet which way is going to move.
As soon as it is rendered, it will collapse to zero
to the approriate edge.
If animation is loaded from the start, this collapsing
takes as long as the animation lasts.  So, for the first time around,
render() will use no animation.