Remove path from AlphaImageLoader; leave file only
AlphaImageLoader will interpret the path relative to
the page, not the css file, so there's no guarantee
that any path in this file will point to the proper
directory according to the implementation.  So remove
the path for byte savings.

This AIL behavior is already documented in the known
issues on the ColorPicker's user guide.  I'll follow
Satyen's advice and call it out more explicitly in the
examples.

Ticket #2528822 (changed from location: code to docs)