Added support to remove shadow when flipped from true to false.
This was missing in general, but also fixes the use case where shadow
is set to false in the constructor configuration and lazyload is true.

Ideally we should never be setting it to true in this later use case,
but due to the way menu lazyload is configured to handle the
configuration queue, we do. Cleaning this up is beyond the scope of
2.9.0.

To support this fix, pulled some pure private methods related to
shadow handling, out into protected prototype methods (to allow
for easy customization, and unsubscription)

Fixes #2529112, #2529266

Tested the test pages for both bugs on Chrome and IE6. Also
ran through some of the menu examples.

NOTE: The changes to contextmenu.js are only whitespace changes